refactors witness and mailbox functionality into seperate packages
