+++ src/keri/app/cli/commands/ends/add.py
@@ -8,8 +8,8 @@ import argparse
 
 from hio.base import doing
 
-from keri import help, kering
-from keri.app import habbing, grouping, indirecting, forwarding
+from keri import help, kering, mailbox
+from keri.app import habbing, grouping, forwarding
 from keri.app.agenting import WitnessPublisher
 from keri.app.cli.common import existing
 from keri.app.notifying import Notifier
@@ -65,7 +65,7 @@ class RoleDoer(doing.DoDoer):
         exc = exchanging.Exchanger(hby=self.hby, handlers=[])
         grouping.loadHandlers(exc, mux)
 
-        mbx = indirecting.MailboxDirector(hby=self.hby, topics=["/receipt", "/multisig", "/replay"], exc=exc)
+        mbx = mailbox.Director(hby=self.hby, topics=["/receipt", "/multisig", "/replay"], exc=exc)
 
         if self.hab is None:
             raise kering.ConfigurationError(f"unknown alias={alias}")
