+++ src/keri/app/cli/commands/mailbox/add.py
@@ -5,6 +5,7 @@ keri.kli.commands module
 
 """
 import argparse
+import warnings
 
 from hio.base import doing
 from hio.help import Hict
@@ -15,6 +16,14 @@ from keri.app.agenting import httpClient, WitnessPublisher
 from keri.app.cli.common import existing
 
 logger = help.ogler.getLogger()
+
+warnings.warn(
+    "Mailbox commands will be removed in a future release. "
+    "Functionality has been moved to its own repository: https://github.com/keri-foundation/mailbox",
+    DeprecationWarning,
+    stacklevel=2,
+)
+
 
 parser = argparse.ArgumentParser(description='Add mailbox role')
 parser.set_defaults(handler=lambda args: add(args),
