+++ src/keri/app/cli/commands/multisig/interact.py
@@ -5,12 +5,12 @@ keri.kli.commands.multisig module
 """
 
 import argparse
-from ordered_set import OrderedSet as oset
 
 from hio.base import doing
+from ordered_set import OrderedSet as oset
 
-from keri import help, kering
-from keri.app import grouping, indirecting, habbing, forwarding
+from keri import help, kering, mailbox
+from keri.app import grouping, habbing, forwarding
 from keri.app.cli.common import existing, displaying, config
 from keri.app.notifying import Notifier
 from keri.core import coring, serdering
@@ -81,7 +81,7 @@ class GroupMultisigInteract(doing.DoDoer):
         exc = exchanging.Exchanger(hby=self.hby, handlers=[])
         grouping.loadHandlers(exc, mux)
 
-        mbd = indirecting.MailboxDirector(hby=self.hby, topics=['/receipt', '/multisig'], exc=exc)
+        mbd = mailbox.Director(hby=self.hby, topics=['/receipt', '/multisig'], exc=exc)
         self.counselor = grouping.Counselor(hby=self.hby)
 
         doers = [self.hbyDoer, self.postman, mbd, self.counselor]
