+++ src/keri/app/cli/commands/vc/registry/incept.py
@@ -2,8 +2,8 @@ import argparse
 
 from hio.base import doing
 
-from keri import help
-from keri.app import indirecting, habbing, grouping, forwarding
+from keri import help, mailbox
+from keri.app import habbing, grouping, forwarding
 from keri.app.cli.common import existing
 from keri.app.habbing import GroupHab
 from keri.app.notifying import Notifier
@@ -95,7 +95,7 @@ class RegistryInceptor(doing.DoDoer):
         exc = exchanging.Exchanger(hby=self.hby, handlers=[])
         grouping.loadHandlers(exc, mux)
 
-        mbx = indirecting.MailboxDirector(hby=self.hby, topics=["/receipt", "/multisig", "/replay"], exc=exc)
+        mbx = mailbox.Director(hby=self.hby, topics=["/receipt", "/multisig", "/replay"], exc=exc)
         self.registrar = credentialing.Registrar(hby=self.hby, rgy=self.rgy, counselor=counselor)
         doers = [self.hbyDoer, counselor, self.registrar, self.postman, mbx]
         self.toRemove = list(doers)
