+++ src/keri/kering.py
@@ -2,14 +2,11 @@
 """
 Generic Constants and Classes
 """
-import sys
 import re
-from collections import namedtuple, deque
+from collections import namedtuple
 from dataclasses import dataclass, astuple
 
-from .help.helping import sceil
-from .help.helping import intToB64, intToB64b, b64ToInt
-
+from .help.helping import intToB64, b64ToInt
 
 MaxON = int("f"*32, 16)  # 256 ** 16 - 1 maximum ordinal number, sequence or first seen etc
