+++ src/keri/witness/__init__.py
@@ -1,0 +1,5 @@
+from .http_end import HttpEnd
+from .receipt_end import ReceiptEnd
+from .setup import setup
+
+__all__ = ['setup']
