+++ src/keri/witness/http_end.py
@@ -1,0 +1,143 @@
+# -*- encoding: utf-8 -*-
+"""
+KERI
+keri.witnessing.http_end module
+
+"""
+
+import falcon
+from hio.help import decking
+
+from .. import help
+from ..app import httping
+from ..core import (eventing, coring)
+from ..core.coring import Ilks
+from ..mailbox import QueryReplyIterable
+
+logger = help.ogler.getLogger()
+
+class HttpEnd:
+    """
+    HTTP handler that accepts and KERI events POSTed as the body of a request with all attachments to
+    the message as a CESR attachment HTTP header.  KEL Messages are processed and added to the database
+    of the provided Habitat.
+
+    This also handles `req`, `exn` and `tel` messages that respond with a KEL replay.
+    """
+
+    TimeoutQNF = 30
+    TimeoutMBX = 5
+
+    def __init__(self, rxbs=None, mbx=None, qrycues=None):
+        """
+        Create the KEL HTTP server from the Habitat with an optional Falcon App to
+        register the routes with.
+
+        Parameters
+             rxbs (bytearray): output queue of bytes for message processing
+             mbx (Mailboxer): Mailbox storage
+             qrycues (Deck): inbound qry response queues
+
+        """
+        self.rxbs = rxbs if rxbs is not None else bytearray()
+
+        self.mbx = mbx
+        self.qrycues = qrycues if qrycues is not None else decking.Deck()
+
+    def on_post(self, req, rep):
+        """
+        Handles POST for KERI event messages.
+
+        Parameters:
+              req (Request) Falcon HTTP request
+              rep (Response) Falcon HTTP response
+
+        ---
+        summary:  Accept KERI events with attachment headers and parse
+        description:  Accept KERI events with attachment headers and parse.
+        tags:
+           - Events
+        requestBody:
+           required: true
+           content:
+             application/json:
+               schema:
+                 type: object
+                 description: KERI event message
+        responses:
+           200:
+              description: Mailbox query response for server sent events
+           204:
+              description: KEL or EXN event accepted.
+        """
+        if req.method == "OPTIONS":
+            rep.status = falcon.HTTP_200
+            return
+
+        rep.set_header('Cache-Control', "no-cache")
+        rep.set_header('connection', "close")
+
+        cr = httping.parseCesrHttpRequest(req=req)
+        sadder = coring.Sadder(ked=cr.payload, kind=eventing.Kinds.json)
+        msg = bytearray(sadder.raw)
+        msg.extend(cr.attachments.encode("utf-8"))
+
+        self.rxbs.extend(msg)
+
+        if sadder.proto in ("ACDC",):
+            rep.set_header('Content-Type', "application/json")
+            rep.status = falcon.HTTP_204
+        else:
+            ilk = sadder.ked["t"]
+            if ilk in (Ilks.icp, Ilks.rot, Ilks.ixn, Ilks.dip, Ilks.drt, Ilks.exn, Ilks.rpy):
+                rep.set_header('Content-Type', "application/json")
+                rep.status = falcon.HTTP_204
+            elif ilk in (Ilks.vcp, Ilks.vrt, Ilks.iss, Ilks.rev, Ilks.bis, Ilks.brv):
+                rep.set_header('Content-Type', "application/json")
+                rep.status = falcon.HTTP_204
+            elif ilk in (Ilks.qry,):
+                if sadder.ked["r"] in ("mbx",):
+                    rep.set_header('Content-Type', "text/event-stream")
+                    rep.status = falcon.HTTP_200
+                    rep.stream = QueryReplyIterable(mbx=self.mbx, cues=self.qrycues, said=sadder.said)
+                else:
+                    rep.set_header('Content-Type', "application/json")
+                    rep.status = falcon.HTTP_204
+
+    def on_put(self, req, rep):
+        """
+        Handles PUT for KERI mbx event messages.
+
+        Parameters:
+              req (Request) Falcon HTTP request
+              rep (Response) Falcon HTTP response
+
+        ---
+        summary:  Accept KERI events with attachment headers and parse
+        description:  Accept KERI events with attachment headers and parse.
+        tags:
+           - Events
+        requestBody:
+           required: true
+           content:
+             application/json:
+               schema:
+                 type: object
+                 description: KERI event message
+        responses:
+           200:
+              description: Mailbox query response for server sent events
+           204:
+              description: KEL or EXN event accepted.
+        """
+        if req.method == "OPTIONS":
+            rep.status = falcon.HTTP_200
+            return
+
+        rep.set_header('Cache-Control', "no-cache")
+        rep.set_header('connection', "close")
+
+        self.rxbs.extend(req.bounded_stream.read())
+
+        rep.set_header('Content-Type', "application/json")
+        rep.status = falcon.HTTP_204
